Fixed issue where keyboard showed return key.
Added a function to PackageTasksFragment.java to allow toggling the keyboard. It takes context, a view, and a mode.
Context is just getContext(), and view is the view you want it to focus on I believe. It can be getView(), or in the case of mSearch.setOnClickListener, mSearchWord(The search box).
Mode is also self explanatory, switch statement, 1 shows the keyboard, 0 hides it, and default hides it
Also changed the search box to be left justified to match with other apps more.